+++ path_tracer.py
@@ -1,248 +1,13 @@
 from PIL import Image
 from optparse import OptionParser
 import numpy as np
-import math
-import glm
-import random
+from materials import *
 from threading import Thread
 from multiprocessing import JoinableQueue
 import sys
 import datetime
+from scenes import *
 
-class Camera:
-    def __init__(self):
-        self.position = glm.vec3(278, 273, -800)
-        self.direction = glm.vec3(0,0,1)
-        self.up = glm.vec3(0,1,0)
-        self.focal_length = 0.035
-        self.width = 0.025
-        self.height = 0.025
-
-class Ray:
-    def __init__(self, origin=glm.vec3(0), direction=glm.vec3(0,0,1)):
-        self.origin = origin
-        self.direction = direction
-
-class Material:
-    # The bidirectional reflectance distribution function which returns
-    # the percentage of light reflected in direction w_o from direction w_i
-    # at location p with normal N.
-    def brdf(self, w_i, w_o, n, p):
-        return glm.vec3(0)
-
-    # Material subclasses implement their own sample functions which
-    # return a new ray along with the pdf of that ray.
-    def sample_ray(self, ray, N, P):
-        return (None, 0)
-
-class DiffuseMaterial(Material):
-    def __init__(self, diffuse=glm.vec3(1), specular=glm.vec3(0), emission = glm.vec3(0), percent=1):
-        self.diffuse_color = diffuse
-        self.specular_color = specular
-        self.emissive_color = emission
-        self.diffuse_percent = percent
-
-    # The diffuse BRDF is equal in all directions with a cosine falloff.
-    def brdf(self, w_i, w_o, n, p):
-        return self.diffuse_color / math.pi
-
-    # Returns a vector in the hemisphere around N with a cosine weighted
-    # distribution to avoid clumping along with the pdf of cos(theta) / Pi.
-    def sample_ray(self, ray, N, P):
-        xi1 = random.uniform(0, 1)
-        xi2 = random.uniform(0, 1)
-    
-        theta = math.acos(math.sqrt(float(1.)-xi1))
-        phi = 2.0*math.pi*xi2
-    
-        xs = math.sin(theta) * math.cos(phi)
-        ys = math.cos(theta)
-        zs = math.sin(theta) * math.sin(phi)
-    
-        y = glm.vec3(N)
-        h = glm.vec3(y)
-        if math.fabs(h.x) <= math.fabs(h.y) and math.fabs(h.x) <= math.fabs(h.z):
-            h.x = 1.0
-        elif math.fabs(h.y) <= math.fabs(h.x) and math.fabs(h.y) <= math.fabs(h.z):
-            h.y= 1.0
-        else:
-            h.z = 1.0
-
-        x = glm.normalize(glm.cross(h,y))
-        z = glm.normalize(glm.cross(x,y))
-
-        new_dir = glm.normalize(xs*x + ys*y + zs*z)
-        assert glm.dot(new_dir, N)
-
-        new_ray = Ray()
-        new_ray.direction = new_dir
-
-        # Add a small epsilon to the new ray starting point to prevent self-intersection
-        # with the object its already on.
-        new_ray.origin = P + 0.001 * new_dir
-        return (new_ray, glm.dot(new_dir, N) / math.pi)
-
-
-class MirrorMaterial(Material):
-    def __init__(self):
-        return
-
-class BoundingBox:
-    def __init__(self, vertices):
-        min_pos = glm.vec3(100000000)
-        max_pos = glm.vec3(-100000000)
-        for v in vertices:
-            if v.x < min_pos.x: min_pos.x = v.x
-            if v.y < min_pos.y: min_pos.y = v.y
-            if v.z < min_pos.z: min_pos.z = v.z
-            if v.x > max_pos.x: max_pos.x = v.x
-            if v.y > max_pos.y: max_pos.y = v.y
-            if v.z > max_pos.z: max_pos.z = v.z
-        self.min = min_pos
-        self.max = max_pos
-
-    def intersect(self, ray):
-        tmin = (self.min.x - ray.origin.x) / (ray.direction.x + 0.0001)
-        tmax = (self.max.x - ray.origin.x) / (ray.direction.x + 0.0001)
- 
-        if tmin > tmax:  tmin, tmax = tmax, tmin
- 
-        tymin = (self.min.y - ray.origin.y) / (ray.direction.y + 0.0001)
-        tymax = (self.max.y - ray.origin.y) / (ray.direction.y + 0.0001)
- 
-        if tymin > tymax: tymin, tymax = tymax, tymin 
- 
-        if (tmin > tymax) or (tymin > tmax):
-            return False 
- 
-        if tymin > tmin:
-            tmin = tymin 
- 
-        if tymax < tmax:
-            tmax = tymax 
- 
-        tzmin = (self.min.z - ray.origin.z) / (ray.direction.z + 0.0001)
-        tzmax = (self.max.z - ray.origin.z) / (ray.direction.z + 0.0001)
- 
-        if tzmin > tzmax: tzmin, tzmax = tzmax, tzmin
- 
-        if (tmin > tzmax) or (tzmin > tmax):
-            return False
- 
-        if tzmin > tmin:
-            tmin = tzmin
- 
-        if tzmax < tmax:
-            tmax = tzmax 
- 
-        return True; 
-
-class Shape:
-    def __init__(self):
-        return
-
-    # Intersects a ray with a shape - subclasses implement their own
-    # intersection routines.
-    def intersect(self, ray):
-        return -1.0
-
-class Mesh(Shape):
-    def __init__(self, vertices, indices, material):
-        self.vertices = vertices
-        self.indices = indices
-        self.material = material
-        self.bbox = BoundingBox(vertices)
-
-    def intersect(self, ray):
-        if not self.bbox.intersect(ray):
-            return (-1, None, None)
-
-        closest_hit = 1000000000.0
-        final_v0 = 0 
-        final_v1 = 0 
-        final_v2 = 0
-        for i in range(0, len(self.indices), 3):
-            v0 = self.vertices[self.indices[i]]
-            v1 = self.vertices[self.indices[i+1]]
-            v2 = self.vertices[self.indices[i+2]]
-            curr_hit = Mesh.triangle_intersect(ray, v0, v1, v2)
-            if curr_hit > 0.0 and curr_hit < closest_hit:
-                closest_hit = curr_hit
-                final_v0 = v0 
-                final_v1 = v1 
-                final_v2 = v2
-            
-        if closest_hit != 1000000000.0:
-            return (closest_hit, Mesh.triangle_normal(final_v0, final_v1, final_v2), self.material)
-        else:
-            return (-1, None, None)
-        
-
-    def triangle_normal(v0, v1, v2):
-        return glm.normalize(glm.cross(v0-v1, v0-v2))
-
-    def triangle_intersect(r, v0, v1, v2):
-        v0v1 = v1 - v0
-        v0v2 = v2 - v0
-        pvec = glm.cross(r.direction, v0v2)
-
-        det = glm.dot(v0v1, pvec)
-
-        if det < 0.000001:
-            return -1.0
-
-        invDet = 1.0 / det
-        tvec = r.origin - v0
-        u = glm.dot(tvec, pvec) * invDet
-
-        if u < 0 or u > 1:
-            return float(-1.0)
-
-        qvec = glm.cross(tvec, v0v1)
-        v = glm.dot(r.direction, qvec) * invDet
-
-        if v < 0 or u + v > 1:
-            return -1.0
-        return glm.dot(v0v2,qvec) * invDet
-
-class Rectangle(Mesh):
-    def __init__(self, v0, v1, v2, v3, material):
-        Mesh.__init__(self, [v0, v1, v2, v3], [0, 1, 2, 0, 2, 3], material)
-
-class Sphere(Shape):
-    def __init__(self, radius=1.0, center=glm.vec3(0), diffuse=glm.vec3(1), specular=glm.vec3(0), percent=1.0):
-        self.radius = radius
-        self.center = center
-        self.material = Material(diffuse, specular, percent)
-
-    # Returns the normal vector at a particular point.
-    def normal(point):
-        return glm.normalize(point - self.center)
-
-    def intersect(self, ray):
-        op = self.center-ray.origin 
-        eps = 1e-4
-        b = glm.dot(op, ray.direction)
-        det = b*b - glm.dot(op, op) + self.radius*self.radius
-        if det < 0:
-            return -1.0
-        else:
-            det = math.sqrt(det)
-        t = b - det
-
-        if b - det > eps:
-            t = b - det 
-        elif b + det > eps:
-            t = b + det
-        else:
-            t = -1.0
-
-        return (t, self.normal(ray.origin + t*ray.direction), self.material) if t != -1.0 else (-1, None, None)
-
-class Scene:
-    def __init__(self, camera, shapes):
-        self.camera = camera
-        self.shapes = shapes
 
 def find_hit(ray, shapes):
     closest_hit = 10000000000
@@ -254,7 +19,6 @@ def find_hit(ray, shapes):
             final_hit = hit
 
     return final_hit
-
 
 def trace_path(ray, shapes, depth, max_depth):
     if depth > max_depth:
@@ -352,119 +116,7 @@ def main():
     
     pixel_data = np.zeros((y_res, x_res, 3))
 
-    camera = Camera()
-
-    shapes = [
-        # Floor - White
-        Rectangle(glm.vec3(552.8, 0.0, 0.0),
-                  glm.vec3(0),
-                  glm.vec3(0,0, 559.2),
-                  glm.vec3(549.6, 0.0, 559.2),
-                  DiffuseMaterial(diffuse = glm.vec3(0.9))),
-
-        # Left wall - Red
-        Rectangle(glm.vec3(552.8,   0.0,   0.0),
-                  glm.vec3(549.6,   0.0, 559.2),
-                  glm.vec3(556.0, 548.8, 559.2),
-                  glm.vec3(556.0, 548.8,   0.0),
-                  DiffuseMaterial(diffuse = glm.vec3(0.9,0.05,0.05))),
-
-        # Right wall - Green
-        Rectangle(glm.vec3(0.0,  0.0, 559.2),
-                  glm.vec3(0.0,   0.0,   0.0),
-                  glm.vec3(0.0, 548.8,   0.0),
-                  glm.vec3(0.0, 548.8, 559.2),
-                  DiffuseMaterial(diffuse = glm.vec3(0.05,0.9,0.05))),
-
-        # Back wall - White
-        Rectangle(glm.vec3(549.6, 0.0, 559.2),
-                  glm.vec3(0.0,  0.0, 559.2),
-                  glm.vec3(0.0, 548.8, 559.2),
-                  glm.vec3(556.0, 548.8, 559.2),
-                  DiffuseMaterial(diffuse = glm.vec3(0.9))),
-
-        # Ceiling - White
-        Rectangle(glm.vec3(556.0, 548.8, 0.0),
-                  glm.vec3(556.0, 548.8, 559.2),
-                  glm.vec3(0.0, 548.8, 559.2),
-                  glm.vec3(0.0, 548.8, 0.0),
-                  DiffuseMaterial(diffuse = glm.vec3(0.9))),
-
-        # Light
-        Rectangle(glm.vec3(343.0, 548.75, 227.0),
-                  glm.vec3(343.0, 548.75, 332.0),
-                  glm.vec3(213.0, 548.75, 332.0),
-                  glm.vec3(213.0, 548.75, 227.0),
-                  DiffuseMaterial(diffuse = glm.vec3(0), emission = glm.vec3(50, 50, 50))),
-
-        # Tall box - White
-        Mesh([glm.vec3(423.0, 330.0, 247.0),
-              glm.vec3(265.0, 330.0, 296.0),
-              glm.vec3(314.0, 330.0, 456.0),
-              glm.vec3(472.0, 330.0, 406.0),
-
-              glm.vec3(423.0,   0.0, 247.0),
-              glm.vec3(423.0, 330.0, 247.0),
-              glm.vec3(472.0, 330.0, 406.0),
-              glm.vec3(472.0,   0.0, 406.0),
-
-              glm.vec3(472.0,   0.0, 406.0),
-              glm.vec3(472.0, 330.0, 406.0),
-              glm.vec3(314.0, 330.0, 456.0),
-              glm.vec3(314.0,   0.0, 456.0),
-
-              glm.vec3(314.0,   0.0, 456.0),
-              glm.vec3(314.0, 330.0, 456.0),
-              glm.vec3(265.0, 330.0, 296.0),
-              glm.vec3(265.0,   0.0, 296.0),
-
-              glm.vec3(265.0,   0.0, 296.0),
-              glm.vec3(265.0, 330.0, 296.0),
-              glm.vec3(423.0, 330.0, 247.0),
-              glm.vec3(423.0,   0.0, 247.0)],
-
-              [0, 1, 2, 0, 2, 3,
-               4, 5, 6, 4, 6, 7,
-               8, 9, 10, 8, 10, 11,
-               12, 13, 14, 12, 14, 15,
-               16, 17, 18, 16, 18, 19],
-
-               DiffuseMaterial(diffuse = glm.vec3(0.7))),
-
-        # Short box - White
-        Mesh([glm.vec3(130.0, 165.0, 65.0),
-              glm.vec3(82.0, 165.0, 225.0),
-              glm.vec3(240.0, 165.0, 272.0),
-              glm.vec3(290.0, 165.0, 114.0),
-
-              glm.vec3(290.0, 0.0, 114.0),
-              glm.vec3(290.0, 165.0, 114.0),
-              glm.vec3(240.0, 165.0, 272.0),
-              glm.vec3(240.0,   0.0, 272.0),
-
-              glm.vec3(130.0,   0.0,  65.0),
-              glm.vec3(130.0, 165.0,  65.0),
-              glm.vec3(290.0, 165.0, 114.0),
-              glm.vec3(290.0,   0.0, 114.0),
-
-              glm.vec3(82.0,   0.0, 225.0),
-              glm.vec3(82.0, 165.0, 225.0),
-              glm.vec3(130.0, 165.0,  65.0),
-              glm.vec3(130.0,   0.0,  65.0),
-
-              glm.vec3(240.0,   0.0, 272.0),
-              glm.vec3(240.0, 165.0, 272.0),
-              glm.vec3(82.0, 165.0, 225.0),
-              glm.vec3(82.0,   0.0, 225.0)],
-
-             [0, 1, 2, 0, 2, 3,
-              4, 5, 6, 4, 6, 7,
-              8, 9, 10, 8, 10, 11,
-              12, 13, 14, 12, 14, 15,
-              16, 17, 18, 16, 18, 19],
-
-              DiffuseMaterial(diffuse = glm.vec3(0.7)))
-    ]
+    scene = CornellBox()
 
     q = JoinableQueue()
 
@@ -477,8 +129,8 @@ def main():
             x,y = pixel
             color = glm.vec3(0)
             for s in range(samples):
-                ray = generate_ray(x, y, x_res, y_res, camera)
-                color += trace_path(ray, shapes, 0, 7)
+                ray = generate_ray(x, y, x_res, y_res, scene.camera)
+                color += trace_path(ray, scene.shapes, 0, 7)
             color /= samples
             pixel_data[y, x, 0] = color.x
             pixel_data[y, x, 1] = color.y
